/geral_mp4 now correctly traverses a playlist
The playlist is still hardcoded in the view.

Each movie source has to have the *full* URL (e.g.
http://localhost:3000/mp4/wut.mp4), because videojs-playlist uses the
full source value to obtain the current movie's index.

We're looping each video (with `loop` option on the video tag) and only
going to the next video when the playlist item `duration` expires.

We disabled the spinner and set the poster (the still image shown before
the movie starts) as a white image. The script reloads the page after
playing the playlist 3 times (to fetch new spots).